script/Windows/98t爬取/sehuatang-server/requirements.txt
Flask>=3.0.0
Flask-Login>=0.6.0
bcrypt>=4.0.0
APScheduler>=3.10.0
DrissionPage>=4.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.28.0
python-telegram-bot>=20.0
gunicorn>=21.2.0
